requirements.txt
pydantic-evals==0.3.7
aiohttp==3.12.13
logfire==3.23.0
python-dateutil==2.9.0.post0
